Extract StoryService's content-analysis heuristics into a tested module
storyService.ts was the largest file in the repo: a god class mixing AI
orchestration with a cluster of pure content-analysis heuristics
(extractCharacterNames, extractPlotThreads, analyzeEmotionalTone,
extractThemesFromContent, extractSpicyLevelFromContent, and more) that
have needed several one-off regex bugfixes in recent PRs. The tests for
them already proved the design was wrong, reaching into private methods
via `as any` casts and a hand-redeclared interface just to call a pure
string-to-string function.

Move every state-free method into storyContentAnalysis.ts as plain
exported functions with unchanged behavior (storyService.ts: 2,219 ->
1,682 lines). Add story-content-analysis.test.ts with direct coverage,
including first-ever dedicated tests for extractCharacterNames,
extractPlotThreads, extractChapterTitleAndBody, createContextExcerpt,
getCreatureDisplayName, getSpicyLabel, and formatChapterContent. Clean
up the `as any` reach-ins in the two existing StoryService test files.

Pure extraction, no behavior change.
