+++ api/_lib/services/storyService.ts
@@ -6,12 +6,10 @@ import {
   ChapterContinuationSeam,
   ApiResponse,
   VALIDATION_RULES,
-  SpicyLevel,
   Chapter,
   ChapterFailure,
   CliffhangerType,
-  CreatureType,
-  ThemeType
+  CreatureType
 } from '../types/contracts';
 import { selectRandomAuthorStyles } from '../config/authorStyles';
 import { CliffhangerService, hasIdentifiedCliffhangerType } from './cliffhangerService';
@@ -19,8 +17,24 @@ import { TropeSelection, TropeSubversionService } from './tropeSubversionService
 import { logger, logError, logWarn, logApiError, logInfo, logPerformance, LogContext } from '../utils/logger';
 import { getXaiFastTimeoutMs, getXaiPrimaryTimeoutMs, type XaiReasoningEffort } from '../config/xaiConfig';
 import { XaiTextClient, type XaiTextResponse } from './xaiTextClient';
-import { splitStoryIntoTextBlocks, stripStoryHtmlToText } from '../../../shared/storyTextBlocks';
-import { capAtWordBoundary, tailAtWordBoundary } from '../utils/textExcerpt';
+import {
+  analyzeEmotionalTone,
+  countWords,
+  createContextExcerpt,
+  extractCharacterNames,
+  extractChapterTitleAndBody,
+  extractLastChapterSummary,
+  extractPlotThreads,
+  extractSpicyLevelFromContent,
+  extractThemesFromContent,
+  formatChapterContent,
+  formatStoryContent,
+  generateNextChapterHint,
+  getCreatureDisplayName,
+  getSpicyLabel,
+  stripHtml,
+  stripSpeakerTagsForDisplay
+} from './storyContentAnalysis';
 import {
   UNRECOGNIZED_PARAMETER,
   toLoggableBoolean,
@@ -56,12 +70,6 @@ interface GeneratedChaptersResult {
   aiMetadata?: AiCallMetadata;
 }
 
-/**
- * The longest `nextChapterHint`, in code points. Unchanged from the 200 the
- * `slice(0, 197)` it replaces measured against — the length is not what was
- * wrong with it; the ellipsis is still counted against the same total.
- */
-const NEXT_CHAPTER_HINT_MAX_LENGTH = 200;
 const EXTRA_BATCH_CHAPTER_TIMEOUT_MS = 9000;
 const MOCK_CONTINUATION_TARGET_BODY_WORDS = 450;
 // The prompt boundary's own reading of a theme seed, taken from the shared
@@ -74,9 +82,6 @@ const STORY_LAB_THEME_LABEL_MAX_LENGTH = STORY_BLUEPRINT_LIMITS.maxThemeLabelLen
 const STORY_LAB_THEME_DESCRIPTION_MAX_LENGTH = STORY_BLUEPRINT_LIMITS.maxThemeDescriptionLength;
 const STORY_LAB_CONTEXT_VALUE_MAX_LENGTH = 320;
 const STORY_LAB_NO_GO_CONTENT_MAX_LENGTH = STORY_LAB_CONTEXT_VALUE_MAX_LENGTH;
-// How much of the closing passage the continuation prompt is shown as "what
-// just happened", in words.
-const SUMMARY_WORD_LIMIT = 150;
 
 export class StoryService {
   private readonly xaiClient = new XaiTextClient();
@@ -313,8 +318,8 @@ export class StoryService {
         );
         aiMetadata = this.mergeAiMetadata(aiMetadata, generatedText.aiMetadata);
         const rawChapterContent = generatedText.content;
-        const displayContent = this.stripSpeakerTagsForDisplay(rawChapterContent);
-        const { title, body } = this.extractChapterTitleAndBody(displayContent, chapterNumber);
+        const displayContent = stripSpeakerTagsForDisplay(rawChapterContent);
+        const { title, body } = extractChapterTitleAndBody(displayContent, chapterNumber);
         const chapterContent = body || displayContent;
         const cliffhanger = this.detectCliffhanger(chapterContent);
         const chapter: Chapter = {
@@ -323,11 +328,11 @@ export class StoryService {
           title,
           content: chapterContent,
           rawContent: rawChapterContent,
-          wordCount: this.countWords(chapterContent),
+          wordCount: countWords(chapterContent),
           generatedAt: new Date(),
           hasAudio: false,
           cliffhangerEnding: cliffhanger,
-          nextChapterHint: this.generateNextChapterHint(chapterContent)
+          nextChapterHint: generateNextChapterHint(chapterContent)
         };
 
         chapters.push(chapter);
@@ -455,8 +460,8 @@ export class StoryService {
           );
           aiMetadata = this.mergeAiMetadata(aiMetadata, generatedText.aiMetadata);
           const rawChapterContent = generatedText.content;
-          const displayContent = this.stripSpeakerTagsForDisplay(rawChapterContent);
-          const { title, body } = this.extractChapterTitleAndBody(displayContent, chapterNumber);
+          const displayContent = stripSpeakerTagsForDisplay(rawChapterContent);
+          const { title, body } = extractChapterTitleAndBody(displayContent, chapterNumber);
           const chapterContent = body || displayContent;
           const cliffhangerAnalysis = this.cliffhangerService.analyze(chapterContent, previousCliffhangers);
           lastCliffhangerAnalysis = cliffhangerAnalysis;
@@ -470,11 +475,11 @@ export class StoryService {
             title,
             content: chapterContent,
             rawContent: rawChapterContent,
-            wordCount: this.countWords(chapterContent),
+            wordCount: countWords(chapterContent),
             generatedAt: new Date(),
             hasAudio: false,
             cliffhangerEnding: cliffhangerAnalysis.cliffhangerDetected,
-            nextChapterHint: this.generateNextChapterHint(chapterContent)
+            nextChapterHint: generateNextChapterHint(chapterContent)
           };
 
           chapters.push(chapter);
@@ -519,7 +524,7 @@ export class StoryService {
 
       const firstChapter = chapters[0];
       const lastChapter = chapters[chapters.length - 1];
-      const totalWordCount = this.countWords(aggregatedHtml);
+      const totalWordCount = countWords(aggregatedHtml);
 
       // Create response
       const output: ChapterContinuationSeam['output'] = {
@@ -529,8 +534,8 @@ export class StoryService {
         content: firstChapter.content,
         wordCount: firstChapter.wordCount,
         cliffhangerEnding: Boolean(lastChapter.cliffhangerEnding),
-        themesContinued: this.extractThemesFromContent(aggregatedHtml),
-        spicyLevelMaintained: this.extractSpicyLevelFromContent(aggregatedHtml),
+        themesContinued: extractThemesFromContent(aggregatedHtml),
+        spicyLevelMaintained: extractSpicyLevelFromContent(aggregatedHtml),
         appendedToStory: aggregatedHtml,
         tropeMetadata: sanitizedInput.tropeMetadata,
         cliffhangerAnalysis: lastCliffhangerAnalysis,
@@ -657,7 +662,7 @@ export class StoryService {
       });
 
       return {
-        content: this.formatStoryContent(response.text),
+        content: formatStoryContent(response.text),
         aiMetadata: this.toAiCallMetadata(response)
       };
 
@@ -728,7 +733,7 @@ export class StoryService {
       });
 
       return {
-        content: this.formatChapterContent(response.text),
+        content: formatChapterContent(response.text),
         aiMetadata: this.toAiCallMetadata(response)
       };
 
@@ -1109,9 +1114,9 @@ Your goal: Create episodes that make listeners desperate for "Continue Chapter."
   }
 
   private buildUserPrompt(input: StoryGenerationSeam['input']): string {
-    const creatureName = this.getCreatureDisplayName(input.creature);
+    const creatureName = getCreatureDisplayName(input.creature);
     const themesText = this.formatThemeContext(input);
-    const spicyLabel = this.getSpicyLabel(input.spicyLevel);
+    const spicyLabel = getSpicyLabel(input.spicyLevel);
     const chekovElements = this.generateChekovElements();
     const storyLabContext = this.formatStoryLabContext(input);
 
@@ -1277,7 +1282,7 @@ Plant your Chekhov elements naturally and ensure the moral dilemma occurs at mid
     const basePrompt = this.buildUserPrompt(input);
     const [, ...restLines] = basePrompt.split('\n');
     const contextExcerpt = options.existingContent
-      ? `PREVIOUS CHAPTER EXCERPT (for continuity, do not repeat verbatim):\n${this.createContextExcerpt(options.existingContent)}\n\n`
+      ? `PREVIOUS CHAPTER EXCERPT (for continuity, do not repeat verbatim):\n${createContextExcerpt(options.existingContent)}\n\n`
       : '';
 
     return `Write Chapter ${options.chapterNumber} of ${options.totalChapters} continuing the same supernatural romance saga.
@@ -1293,10 +1298,10 @@ ${contextExcerpt}${restLines.join('\n')}`;
   ): string {
     // Extract intelligent context from previous chapters
     const existingContent = existingContentOverride || input.existingContent;
-    const characterNames = this.extractCharacterNames(existingContent);
-    const lastChapterSummary = this.extractLastChapterSummary(existingContent);
-    const activePlotThreads = this.extractPlotThreads(existingContent);
-    const emotionalTone = this.analyzeEmotionalTone(existingContent);
+    const characterNames = extractCharacterNames(existingContent);
+    const lastChapterSummary = extractLastChapterSummary(existingContent);
+    const activePlotThreads = extractPlotThreads(existingContent);
+    const emotionalTone = analyzeEmotionalTone(existingContent);
     
     const prompt = `Continue this story as Chapter ${chapterNumber}.
 
@@ -1327,7 +1332,7 @@ ${input.userInput ? `CREATIVE DIRECTION: ${input.userInput}` : ''}
 ${this.formatContinuationStoryLabContext(input.generationContext)}
 
 PREVIOUS CHAPTER(S) FOR CONTINUITY:
-${this.createContextExcerpt(existingContent)}
+${createContextExcerpt(existingContent)}
 
 Write 400-600 words for this chapter. Use HTML: <h3> for chapter title, <p> for paragraphs, <em> for emphasis.`;
 
@@ -1358,105 +1363,6 @@ Write 400-600 words for this chapter. Use HTML: <h3> for chapter title, <p> for 
 
     lines.push('- Keep continuation intimacy consensual and do not exceed the original Heat Contract boundary.');
     return lines.join('\n');
-  }
-
-  /**
-   * Extract character names from story content
-   */
-  private extractCharacterNames(content: string): string[] {
-    const speakerMatches = content.match(/\[([^\],]+)(?:,\s*[^\]]+)?\]:/g) || [];
-    const names = speakerMatches
-      .map(match => match.replace(/\[([^\],]+).*/, '$1').trim())
-      .filter(name => name !== 'Narrator');
-    
-    // Deduplicate and return
-    return [...new Set(names)];
-  }
-
-  /**
-   * Extract summary of last chapter/section
-   */
-  private extractLastChapterSummary(content: string): string {
-    // Stories arrive as generator HTML, where a paragraph is a `<p>` element
-    // rather than a run of text between blank lines. Splitting the stripped
-    // text on blank lines therefore found exactly one paragraph — the whole
-    // story — and "the last three paragraphs, truncated to 150 words" became
-    // "the story's opening 150 words". That summary is what the continuation
-    // prompt is told just happened, so a chapter continued from the beginning.
-    const paragraphs = splitStoryIntoTextBlocks(content);
-
-    if (paragraphs.length === 0) return 'Story beginning';
-    
-    // Get last 2-3 paragraphs as summary
-    const lastParagraphs = paragraphs.slice(-3).join(' ');
-
-    // Truncate to ~150 words
-    const words = lastParagraphs.split(/\s+/);
-    const summary = words.slice(0, SUMMARY_WORD_LIMIT).join(' ');
-
-    // Whether the cut happened is a question about the words, not about the
-    // lengths of two strings. Joining on single spaces is itself shortening —
-    // any line break or double space inside the paragraphs comes back as one
-    // character — so comparing lengths reported a truncation for a summary that
-    // holds the whole passage. The marker is what tells the continuation prompt
-    // that the chapter it is being handed stops mid-thought, and the model is
-    // then prompted to resume from a sentence that had in fact already ended.
-    return words.length > SUMMARY_WORD_LIMIT ? summary + '...' : summary;
-  }
-
-  /**
-   * Extract active plot threads and unresolved elements
-   */
-  private extractPlotThreads(content: string): string[] {
-    const threads: string[] = [];
-    const lowerContent = content.toLowerCase();
-
-    // Check for common plot thread indicators
-    if (lowerContent.includes('secret') || lowerContent.includes('mystery')) {
-      threads.push('Unresolved mystery or secret');
-    }
-    if (lowerContent.includes('danger') || lowerContent.includes('threat')) {
-      threads.push('Active threat or danger');
-    }
-    if (lowerContent.includes('forbidden') || lowerContent.includes('impossible')) {
-      threads.push('Forbidden relationship tension');
-    }
-    if (lowerContent.includes('power') || lowerContent.includes('control')) {
-      threads.push('Power dynamics in play');
-    }
-    if (lowerContent.match(/\bwhat\s+(if|would|could)\b/)) {
-      threads.push('Unresolved questions');
-    }
-    
-    return threads.length > 0 ? threads : ['Character development', 'Relationship progression'];
-  }
-
-  /**
-   * Analyze emotional tone of existing content
-   *
-   * `dominan` was a word stem left behind from a substring scan, and every
-   * keyword here is matched as a whole word. Nothing in English is spelled
-   * `dominan`, so the alternative could never fire: the one register the
-   * `intense` tone exists to name — a chapter written about dominance — was
-   * recognised only if it also happened to say `power`, `control`, or
-   * `command`, and a scene that says `dominant` and nothing else was reported
-   * to the continuation prompt as `romantic with building tension`. The
-   * inflections the stem stood for are spelled out instead, which is the same
-   * repair `extractThemesFromContent` made when it moved to whole-word
-   * matching.
-   */
-  private analyzeEmotionalTone(content: string): string {
-    const lowerContent = content.toLowerCase();
-    const tones: string[] = [];
-
-    // Emotional indicators
-    if (lowerContent.match(/\b(desire|passion|want|need|crave)\b/)) tones.push('passionate');
-    if (lowerContent.match(/\b(dark|shadow|danger|fear|threat)\b/)) tones.push('dark/suspenseful');
-    if (lowerContent.match(/\b(tease|playful|smile|grin|laugh)\b/)) tones.push('playful');
-    if (lowerContent.match(/\b(pain|ache|hurt|wound|scar)\b/)) tones.push('angsty');
-    if (lowerContent.match(/\b(power|control|dominant|dominance|dominated|command)\b/)) tones.push('intense');
-
-    return tones.length > 0 ? tones.join(', ') : 'romantic with building tension';
   }
 
   /**
@@ -1605,23 +1511,6 @@ Write 400-600 words for this chapter. Use HTML: <h3> for chapter title, <p> for 
     return 2;
   }
 
-  private extractChapterTitleAndBody(content: string, chapterNumber: number): { title: string; body: string } {
-    const headingMatch = content.match(/<h3[^>]*>(.*?)<\/h3>/i);
-    let title = headingMatch ? this.stripHtml(headingMatch[1]).trim() : '';
-
-    if (title.toLowerCase().startsWith(`chapter ${chapterNumber}`)) {
-      title = title.slice(`chapter ${chapterNumber}`.length).replace(/^\s*:?/, '').trim();
-    }
-
-    if (!title) {
-      title = `Untitled Chapter ${chapterNumber}`;
-    }
-
-    const body = headingMatch ? content.replace(headingMatch[0], '').trim() : content.trim();
-
-    return { title, body };
-  }
-
   private renderChapterForAppend(chapter: Pick<Chapter, 'chapterNumber' | 'title' | 'content'>): string {
     const heading = `<h3>Chapter ${chapter.chapterNumber}: ${chapter.title}</h3>`;
     const sanitizedContent = chapter.content.replace(/^\s*<h3[^>]*>.*?<\/h3>\s*/i, '').trim();
@@ -1646,46 +1535,9 @@ Write 400-600 words for this chapter. Use HTML: <h3> for chapter title, <p> for 
     return `${existing.trim()}\n\n<hr>\n\n${trimmedAddition}`;
   }
 
-  /**
-   * The last sentence of a chapter, as the hint for what comes next.
-   *
-   * Measured and cut in code points. `candidate.slice(0, 197)` counted UTF-16
-   * code units, so a hint whose 197th unit fell between the halves of a
-   * surrogate pair ended on a lone surrogate — and the story this app writes is
-   * one whose prose carries the occasional astral character. The cut also landed
-   * mid-word wherever it landed, which is what `capAtWordBoundary` is for.
-   */
-  private generateNextChapterHint(content: string): string {
-    const text = this.stripHtml(content).replace(/\s+/g, ' ').trim();
-    if (!text) {
-      return '';
-    }
-
-    const sentences = text.split(/(?<=[.!?])\s+/).filter(Boolean);
-    const candidate = (sentences[sentences.length - 1] || text).trim();
-
-    return Array.from(candidate).length > NEXT_CHAPTER_HINT_MAX_LENGTH
-      ? `${capAtWordBoundary(candidate, NEXT_CHAPTER_HINT_MAX_LENGTH - 3)}...`
-      : candidate;
-  }
-
-  /**
-   * The tail of everything written so far, shown to the model as
-   * `PREVIOUS CHAPTER EXCERPT` when it continues the story.
-   *
-   * `text.slice(-maxLength)` cut at the front, in code units, so the excerpt
-   * could begin on the second half of a surrogate pair — and began mid-word
-   * whatever it began on, which is the first thing the model reads.
-   */
-  private createContextExcerpt(html: string, maxLength: number = 1200): string {
-    const text = this.stripHtml(html || '').replace(/\s+/g, ' ').trim();
-
-    return tailAtWordBoundary(text, maxLength);
-  }
-
   private generateMockStory(input: StoryGenerationSeam['input'], _tropeSelection?: TropeSelection): string {
-    const creatureName = this.getCreatureDisplayName(input.creature);
-    const spicyLabel = this.getSpicyLabel(input.spicyLevel);
+    const creatureName = getCreatureDisplayName(input.creature);
+    const spicyLabel = getSpicyLabel(input.spicyLevel);
     const targetBodyWords = Math.max(200, Math.floor(input.wordCount * 0.9));
     const paragraphs = [
       `In the shadowed alleys of Victorian London, Lady Arabella Worthington found herself drawn to the mysterious stranger who haunted her dreams. His eyes, crimson as fresh-spilled wine, held secrets that both terrified and exhilarated her.`,
@@ -1717,7 +1569,7 @@ ${renderBody()}
   }
 
   private generateMockInitialChapter(input: StoryGenerationSeam['input'], chapterNumber: number, targetWordCount: number): string {
-    const creatureName = this.getCreatureDisplayName(input.creature);
+    const creatureName = getCreatureDisplayName(input.creature);
     const baseTitle = chapterNumber === 1
       ? `The ${creatureName}'s Forbidden Passion`
       : `Secrets of the ${creatureName} - Part ${chapterNumber}`;
@@ -1772,7 +1624,7 @@ ${renderBody()}`;
       paragraphs,
       expansionBeats,
       MOCK_CONTINUATION_TARGET_BODY_WORDS,
-      this.countWords(renderBody())
+      countWords(renderBody())
     );
 
     return `<h3>Chapter ${nextNumber}: The Deeper Shadows</h3>
@@ -1784,12 +1636,12 @@ ${renderBody()}`;
     paragraphs: string[],
     expansionBeats: string[],
     targetBodyWords: number,
-    initialWordCount = this.countWords(paragraphs.join(' '))
+    initialWordCount = countWords(paragraphs.join(' '))
   ): void {
     const countedExpansionBeats = expansionBeats
       .map(beat => ({
         beat,
-        wordCount: this.countWords(beat)
+        wordCount: countWords(beat)
       }))
       .filter(({ wordCount }) => wordCount > 0);
     let bodyWordCount = initialWordCount;
@@ -1803,35 +1655,8 @@ ${renderBody()}`;
     }
   }
 
-  private getCreatureDisplayName(creature: string): string {
-    const names: Record<string, string> = {
-      'vampire': 'Vampire',
-      'werewolf': 'Werewolf',
-      'fairy': 'Fairy',
-      'siren': 'Siren',
-      'djinn': 'Djinn',
-      'witch': 'Witch',
-      'dragon': 'Dragon',
-      'demon': 'Demon',
-      'angel': 'Angel',
-      'mermaid': 'Mermaid'
-    };
-    return names[creature] || 'Creature';
-  }
-
-  private getSpicyLabel(level: number): string {
-    const labels = [
-      'Storybook romance',
-      'Warm',
-      'Spicy',
-      'Very spicy',
-      'Inferno'
-    ];
-    return labels[level - 1] || 'Spicy';
-  }
-
   private generateTitle(input: StoryGenerationSeam['input']): string {
-    const creatureName = this.getCreatureDisplayName(input.creature);
+    const creatureName = getCreatureDisplayName(input.creature);
     return `The ${creatureName}'s Forbidden Passion`;
   }
 
@@ -1839,353 +1664,8 @@ ${renderBody()}`;
     return 'The Deeper Shadows';
   }
 
-  /**
-   * Count the words a reader would count.
-   *
-   * The count is reported to the client as `actualWordCount` and drives the
-   * streaming progress percentage, so it has to match the rendered story rather
-   * than the markup. Stripping tags in place merged the words on either side of
-   * every paragraph break into one, which cost one word per boundary — a
-   * chapter of forty `<p>` elements with no whitespace between them reported
-   * thirty-nine fewer words than it has, and `<p>one</p><p>two</p>` reported a
-   * single word.
-   */
-  private countWords(content: string): number {
-    return stripStoryHtmlToText(content).split(/\s+/).filter(word => word.length > 0).length;
-  }
-
   private detectCliffhanger(content: string): boolean {
     return this.cliffhangerService.analyze(content).cliffhangerDetected;
-  }
-
-  /**
-   * Report which of the reader's themes the new chapters carried on.
-   *
-   * The result is returned to the caller as `themesContinued`, which the
-   * contract types as `ThemeType[]` — the same closed set of eighteen ids the
-   * form offers and `VALIDATION_RULES.themes.allowedValues` lists. Two things
-   * kept it from being one:
-   *
-   * - When nothing matched, the answer was `['romance', 'fantasy']`. Neither is
-   *   a theme: no chapter can be generated with either, no theme picker can
-   *   render either, and a caller mapping the ids back to labels gets nothing
-   *   for both. It is also not the honest answer — "no configured theme was
-   *   detected" is — and because a scan this coarse usually matches something,
-   *   the case it fired in was the one where the scan had found nothing to say.
-   * - Six of the eighteen themes had no keywords at all, so `dominance`,
-   *   `submission`, `temptation`, `sin`, `lust`, and `deceit` could never be
-   *   reported however plainly a chapter carried them: a scene naming all six
-   *   came back as `power_dynamics, desire`. `lust` was worse than absent — it
-   *   sat in `desire`'s keyword list, so the word was credited to a theme the
-   *   reader may not have chosen while its own theme stayed unreachable.
-   *
-   * Keying the table by `ThemeType` is what stops the second from returning: a
-   * theme added to the contract without keywords here is now a compile error
-   * rather than a silent blind spot. The declared return type does the same for
-   * the first.
-   *
-   * The scan reads the rendered text rather than the markup, like every other
-   * scanner here — the multi-word keywords (`secret love`, `star-crossed`,
-   * `false promise`) are the ones a welded `door.</p><p>Blood` boundary hides.
-   *
-   * Keywords are matched as whole words rather than as substrings, which is
-   * what makes the six new entries safe to state plainly: `sin` as a substring
-   * is in `rising`, `using`, and `singing`, and `lust` is in `lustre`, so under
-   * the old matching the only way to add those themes would have been to spell
-   * them as something other than their own names. The inflections the substring
-   * form used to pick up for free — `secrets` for `secret`, `powerful` for
-   * `power` — are listed instead. `used` is gone from `manipulation`: an
-   * ordinary "she used the key" is not a story about being used, and it was the
-   * loosest keyword in the table.
-   */
-  private extractThemesFromContent(content: string): ThemeType[] {
-    const lowerContent = this.stripHtml(content).toLowerCase();
-
-    // Ordered as `VALIDATION_RULES.themes.allowedValues` orders them, so the
-    // same chapter always reports the same list in the same order.
-    const themeKeywords: Record<ThemeType, readonly string[]> = {
-      betrayal: ['betrayed', 'betrayal', 'deceived', 'backstabbed', 'treachery', 'double-crossed'],
-      obsession: ['obsessed', 'obsession', 'possessed', 'consumed', 'fixated', 'addicted'],
-      power_dynamics: ['power', 'powers', 'powerful', 'control', 'authority', 'command', 'leverage'],
-      forbidden_love: ['forbidden', 'secret love', 'star-crossed', 'illicit', 'taboo'],
-      revenge: ['revenge', 'vengeance', 'retribution', 'payback', 'avenge', 'avenged'],
-      manipulation: ['manipulated', 'manipulation', 'controlled', 'exploited', 'influenced'],
-      seduction: ['seduced', 'seduction', 'allured', 'enticed', 'charmed', 'coaxed'],
-      dark_secrets: ['secret', 'secrets', 'hidden', 'mysterious', 'concealed', 'buried'],
-      corruption: ['corrupted', 'corruption', 'tainted', 'fallen', 'darkness', 'evil'],
-      dominance: ['dominance', 'dominant', 'dominated', 'dominion', 'mastery'],
-      submission: ['submission', 'submitted', 'submissive', 'yielded', 'knelt', 'obeyed'],
-      jealousy: ['jealous', 'jealousy', 'envious', 'possessive', 'resentful', 'covetous'],
-      temptation: ['tempted', 'temptation', 'tempting', 'lured', 'beckoned'],
-      sin: ['sin', 'sins', 'sinful', 'sinner', 'damnation', 'damned', 'penance'],
-      desire: ['desire', 'desires', 'yearning', 'craving', 'longing', 'wanting'],
-      passion: ['passionate', 'passion', 'intense', 'burning', 'fiery', 'ardent'],
-      lust: ['lust', 'lustful', 'lusted', 'carnal', 'ravenous'],
-      deceit: ['deceit', 'deceitful', 'lied', 'lying', 'false promise']
-    };
-
-    const detectedThemes: ThemeType[] = [];
-    for (const [theme, keywords] of Object.entries(themeKeywords) as Array<[ThemeType, readonly string[]]>) {
-      if (keywords.some(keyword => containsWholeWord(lowerContent, keyword))) {
-        detectedThemes.push(theme);
-      }
-    }
-
-    return detectedThemes;
-  }
-
-  /**
-   * The spice level a batch of chapters actually reads at.
-   *
-   * The answer travels: it is `spicyLevelMaintained` on the continuation
-   * response, and `buildContinuationPayload` writes it into a new story's
-   * `StorySummary.spicyLevel` — the level the project is then stored and
-   * reopened under. So a misread here is not a cosmetic number beside the
-   * prose; it is what the library says the story is.
-   *
-   * Its sibling `extractThemesFromContent` directly above scans the rendered
-   * text with whole-word matching. This one did neither, and the two failures
-   * compound:
-   *
-   * - **It scanned the markup.** Every other scanner here — the cliffhanger
-   *   service, the image service, the continuity extractor, the story-quality
-   *   heuristics, and `extractThemesFromContent` beside it — reads
-   *   `stripStoryHtmlToText` first, so that what is measured is what the
-   *   reader sees: block tags become paragraph breaks and the entities the
-   *   generator writes are decoded. Undecoded, `intense&nbsp;passion` is not
-   *   the phrase `intense passion` at all, and the one multi-word keyword the
-   *   scan has matched nothing whenever the model spaced it that way.
-   * - **It matched substrings.** `heart` is inside `hearth`, `love` inside
-   *   `glove` and `clover`, `touch` inside `untouched`, `heat` inside
-   *   `sheath` and `wheat`, and `climax` inside `anticlimax` — so a chaste
-   *   scene at a hearth in wool gloves was filed as level 2, a chapter that
-   *   ends `he left her untouched` as level 3, and one that calls a duel an
-   *   anticlimax as level 5, the maximum, on the strength of a word that says
-   *   the opposite. The story then reopens at that level, and the reader who
-   *   set the dial themselves is the one contradicted.
-   *
-   * Keywords are matched as whole words for the same reason
-   * `extractThemesFromContent` was changed to: it is the only way to state
-   * `heat` or `love` as itself. The inflections the substring form picked up
-   * for free — `kissed` for `kiss`, `caressing` for `caress`, `desired` for
-   * `desire` — are listed instead, so the repair does not quietly cost the
-   * scan the matches it did get right. What it does not carry over is the rest
-   * of what the substrings caught: `lovely` is not `love`, `gentleman` is not
-   * `gentle`, and `hearth` is not `heart`. Those are the defect, not coverage.
-   */
-  private extractSpicyLevelFromContent(content: string): SpicyLevel {
-    const lowerContent = this.stripHtml(content).toLowerCase();
-
-    // Level 5 - Very Explicit
-    const level5Keywords = [
-      'explicit', 'explicitly', 'graphic', 'graphically', 'intense passion',
-      'climax', 'climaxed', 'climaxes', 'ecstasy'
-    ];
-    if (level5Keywords.some(keyword => containsWholeWord(lowerContent, keyword))) {
-      return 5 as SpicyLevel;
-    }
-
-    // Level 4 - Passionate
-    const level4Keywords = [
-      'passionate', 'passionately', 'breathless', 'breathlessly',
-      'desire', 'desires', 'desired', 'yearning', 'heat', 'heated'
-    ];
-    if (level4Keywords.some(keyword => containsWholeWord(lowerContent, keyword))) {
-      return 4 as SpicyLevel;
-    }
-
-    // Level 3 - Romantic with Heat
-    const level3Keywords = [
-      'kiss', 'kissed', 'kisses', 'kissing',
-      'embrace', 'embraced', 'embraces', 'embracing',
-      'caress', 'caressed', 'caresses', 'caressing',
-      'touch', 'touched', 'touches', 'touching',
-      'intimate', 'intimately'
-    ];
-    if (level3Keywords.some(keyword => containsWholeWord(lowerContent, keyword))) {
-      return 3 as SpicyLevel;
-    }
-
-    // Level 2 - Sweet Romance
-    const level2Keywords = [
-      'love', 'loved', 'loves', 'lover', 'lovers', 'loving',
-      'affection', 'affections', 'affectionate',
-      'tender', 'tenderly', 'tenderness',
-      'gentle', 'gently', 'heart', 'hearts'
-    ];
-    if (level2Keywords.some(keyword => containsWholeWord(lowerContent, keyword))) {
-      return 2 as SpicyLevel;
-    }
-
-    // Default to Level 1 - Mild
-    return 1 as SpicyLevel;
-  }
-
-  private formatStoryContent(content: string): string {
-    // Enhanced formatting for better readability
-    let formatted = content;
-
-    // If no HTML formatting exists, apply smart formatting
-    if (!content.includes('<h3>') && !content.includes('<p>')) {
-      // Extract title if present (first line typically).
-      //
-      // Only the blank lines *before* the title are dropped. Dropping all of
-      // them — `split('\n').filter(line => line.trim())` — took out the very
-      // separators the paragraph split below looks for, so rejoining the
-      // remainder produced a body with no blank line left anywhere in it and
-      // `split('\n\n')` returned the whole story as one block. Every paragraph
-      // the model wrote was then welded into a single `<p>`, and only for a
-      // story that opens with a title line: the same story without one kept its
-      // paragraphs, because that branch never touched the lines. This is the
-      // path a plain-text answer from the provider takes on its way to the
-      // reader, so what the reader saw was the chapter as one unbroken wall.
-      const lines = content.split('\n');
-      const titleIndex = lines.findIndex(line => line.trim());
-      const firstLine = titleIndex === -1 ? undefined : lines[titleIndex]?.trim();
-
-      // Check if first line looks like a title (short, no punctuation except colon)
-      const isTitle = firstLine && firstLine.length < 80 && !firstLine.endsWith('.') && !firstLine.startsWith('[');
-
-      if (isTitle) {
-        formatted = `<h3>${firstLine}</h3>\n\n` + lines.slice(titleIndex + 1).join('\n');
-      }
-
-      // Split into paragraphs based on multiple newlines or speaker changes
-      formatted = formatted
-        .replace(/\n\s*\n/g, '\n\n') // Normalize line breaks
-        .split('\n\n')
-        .filter(para => para.trim())
-        .map(para => para.trim())
-        .map(para => {
-          // Skip if already has HTML tags
-          if (para.includes('<')) return para;
-          
-          // Wrap in paragraph tags
-          return `<p>${para}</p>`;
-        })
-        .join('\n\n');
-    }
-
-    return formatted;
-  }
-
-  private formatChapterContent(content: string): string {
-    // Enhanced chapter formatting to match story formatting
-    let formatted = content;
-
-    // If no HTML formatting exists, apply smart formatting
-    if (!content.includes('<h3>') && !content.includes('<p>')) {
-      // Split into paragraphs based on multiple newlines
-      formatted = formatted
-        .replace(/\n\s*\n/g, '\n\n') // Normalize line breaks
-        .split('\n\n')
-        .filter(para => para.trim())
-        .map(para => para.trim())
-        .map(para => {
-          // Skip if already has HTML tags
-          if (para.includes('<')) return para;
-          
-          // Wrap in paragraph tags
-          return `<p>${para}</p>`;
-        })
-        .join('\n\n');
-    }
-
-    return formatted;
-  }
-
-  /**
-   * Reduce story markup to the text a reader sees.
-   *
-   * Deleting the tags and nothing else closed the gap they held open, so the
-   * last word of one paragraph and the first word of the next were read as one
-   * token: `<p>She opened the door.</p><p>Blood pooled…</p>` became
-   * `door.Blood`. Every caller here is looking for something the reader can
-   * point at — a chapter title, a summary of what just happened, the sentence a
-   * continuation has to follow on from — and each of them was handed welded
-   * text instead. Sentence splitting suffered worst: with no space after the
-   * full stop there was nothing for `/(?<=[.!?])\s+/` to split on, so the whole
-   * chapter came back as its own final sentence.
-   */
-  private stripHtml(content: string): string {
-    return stripStoryHtmlToText(content);
-  }
-
-  private stripSpeakerTagsForDisplay(content: string): string {
-    // Enhanced speaker tag removal with better text formatting
-    let displayContent = content;
-
-    // Remove speaker tags but preserve structure
-    displayContent = displayContent
-      .replace(/\[([^\]]+?)\]:\s*/g, '') // Remove speaker tags like [Narrator]: [Character, emotion]:
-      .replace(/\n\s*\n/g, '\n\n') // Normalize multiple newlines
-      .trim();
-
-    // Smart paragraph creation based on content structure.
-    //
-    // The blank lines have to survive the split. Filtering them out here left
-    // the branch below — the one that reads a blank line as the paragraph break
-    // it is — unreachable, so the only breaks this method could ever make were
-    // the ones the dialogue and narrative-shift heuristics guessed at. A model
-    // that separated its paragraphs the ordinary way, with a blank line and no
-    // opening quote or `Suddenly` to give itself away, had every one of them
-    // welded into a single `<p>`: the reader was shown the chapter as one
-    // unbroken block, and the paragraph structure the generator had actually
-    // written was thrown away before anything downstream could read it.
-    const lines = displayContent.split('\n');
-    const paragraphs = [];
-    let currentParagraph = '';
-
-    for (const line of lines) {
-      const trimmedLine = line.trim();
-
-      // Empty line indicates paragraph break
-      if (!trimmedLine) {
-        if (currentParagraph) {
-          paragraphs.push(currentParagraph.trim());
-          currentParagraph = '';
-        }
-        continue;
-      }
-
-      // Start new paragraph for dialogue or narrative shifts
-      const isDialogue = trimmedLine.startsWith('"') || trimmedLine.includes('"');
-      const isNarrativeShift = trimmedLine.length < 50 && (
-        trimmedLine.includes('Later') || 
-        trimmedLine.includes('Meanwhile') || 
-        trimmedLine.includes('Suddenly') ||
-        trimmedLine.includes('Then') ||
-        /^(The|As|But|However|Still)/i.test(trimmedLine)
-      );
-
-      if (currentParagraph && (isNarrativeShift || (isDialogue && !currentParagraph.includes('"')))) {
-        paragraphs.push(currentParagraph.trim());
-        currentParagraph = trimmedLine;
-      } else {
-        currentParagraph += (currentParagraph ? ' ' : '') + trimmedLine;
-      }
-    }
-
-    // Add final paragraph
-    if (currentParagraph) {
-      paragraphs.push(currentParagraph.trim());
-    }
-
-    // Format paragraphs with proper HTML
-    const formattedParagraphs = paragraphs
-      .filter(para => para.length > 0)
-      .map(para => {
-        // Clean up any extra spacing
-        para = para.replace(/\s+/g, ' ').trim();
-        
-        // Wrap in paragraph tags if not already formatted
-        if (!para.startsWith('<') && !para.includes('<p>')) {
-          return `<p>${para}</p>`;
-        }
-        return para;
-      });
-
-    return formattedParagraphs.join('\n\n');
   }
 
   private generateStoryId(): string {
@@ -2199,21 +1679,4 @@ ${renderBody()}`;
   private generateRequestId(): string {
     return `req_${randomUUID()}`;
   }
-}
-
-/**
- * Whether `text` contains `keyword` as a whole word or whole phrase.
- *
- * Both sides are already lowercased by the caller. The `\b` at each end is what
- * separates a theme keyword from the longer word it happens to sit inside, and
- * a hyphenated keyword such as `star-crossed` is unaffected: `-` is a
- * non-word character, so the boundaries fall at the ends of the phrase rather
- * than around each half.
- */
-function containsWholeWord(text: string, keyword: string): boolean {
-  return new RegExp(String.raw`\b${escapeRegExp(keyword)}\b`).test(text);
-}
-
-function escapeRegExp(value: string): string {
-  return value.replace(/[.*+?^${}()|[\]\\]/g, String.raw`\$&`);
 }
