+++ tests/story-service-prompt-guards.test.ts
@@ -3,6 +3,7 @@
 
 import { StoryService } from '../api/_lib/services/storyService';
 import type { StoryGenerationSeam } from '../api/_lib/types/contracts';
+import { analyzeEmotionalTone } from '../api/_lib/services/storyContentAnalysis';
 
 function assert(condition: unknown, message: string): asserts condition {
   if (!condition) {
@@ -13,7 +14,6 @@ function assert(condition: unknown, message: string): asserts condition {
 const service = new StoryService() as unknown as {
   formatThemeContext(input: StoryGenerationSeam['input']): string;
   formatStoryLabContext(input: StoryGenerationSeam['input']): string;
-  analyzeEmotionalTone(content: string): string;
 };
 
 const longLabel = 'L'.repeat(120);
@@ -74,25 +74,25 @@ assert(!storyLabContext.includes('undefined'), 'malformed prompt context should 
 // chapter written about dominance and nothing else was described to the model
 // as `romantic with building tension`.
 const dominanceChapter = '<p>She named the terms. He was dominant in every way that counted.</p>';
-const dominanceTone = service.analyzeEmotionalTone(dominanceChapter);
+const dominanceTone = analyzeEmotionalTone(dominanceChapter);
 assert(dominanceTone.includes('intense'), 'a chapter about dominance should read as an intense register');
 assert(
-  service.analyzeEmotionalTone('<p>Her dominance was not in question.</p>').includes('intense'),
+  analyzeEmotionalTone('<p>Her dominance was not in question.</p>').includes('intense'),
   'the dominance inflection should be recognised too'
 );
 assert(
-  service.analyzeEmotionalTone('<p>A quiet supper by the window.</p>') === 'romantic with building tension',
+  analyzeEmotionalTone('<p>A quiet supper by the window.</p>') === 'romantic with building tension',
   'a chapter carrying none of the registers should still fall back honestly'
 );
 
 // The registers that already worked keep working: this repair spells out one
 // alternative, it does not loosen the whole-word matching around it.
 assert(
-  service.analyzeEmotionalTone('<p>He took control of the room.</p>').includes('intense'),
+  analyzeEmotionalTone('<p>He took control of the room.</p>').includes('intense'),
   'the registers that already matched should be unchanged'
 );
 assert(
-  !service.analyzeEmotionalTone('<p>The predominant colour was red.</p>').includes('intense'),
+  !analyzeEmotionalTone('<p>The predominant colour was red.</p>').includes('intense'),
   'a word that merely contains a keyword should still not match'
 );
